refactor(alerter): replace metric switch statements with registry
Introduce metricRegistry map that maps metric names to SQL queries
and scan types. GetLatestMetricValues and GetHistoricalMetricValues
now perform simple registry lookups instead of massive switch
statements.

Part of #75.
